+++ src/ITensorChainRules/zygoterules.jl
@@ -11,3 +11,16 @@ using ZygoteRules: @adjoint
   end
   return y, adjoint_pullback
 end
+
+## XXX: raise issue about `tr` being too generically
+## defined in ChainRules
+##
+## using Zygote
+## 
+## # Needed because by default it was calling the generic
+## # rrule for `tr` inside ChainRules
+## function rrule(::typeof(tr), x::ITensor; kwargs...)
+##   y, tr_pullback_zygote = pullback(ITensors._tr, x; kwargs...)
+##   tr_pullback(ȳ) = (NoTangent(), tr_pullback_zygote(ȳ)...)
+##   return y, tr_pullback
+## end