+++ src/Ops/Ops.jl
@@ -2,9 +2,9 @@ module Ops
 
 using ..LazyApply
 
-import Base: +, -, *, /, exp, show, adjoint
+import Base: ==, +, -, *, /, convert, exp, show, adjoint, isless, hash
 
-export Op, sites, params, Applied, expand
+export Op, OpSum, which_op, site, sites, params, Applied, expand
 
 #####################################################################################
 # General functionality
@@ -17,42 +17,125 @@ export Op, sites, params, Applied, expand
 # (1, "X", 1, 2, "Y", 2, "Z", 4)
 # 
 # julia> split(x -> x isa AbstractString, t)
-# ((1,), ("X", 1, 2), ("Y", 2), ("Z", 4))
+# [(1,), ("X", 1, 2), ("Y", 2), ("Z", 4)]
 # 
 function split(f, t::Tuple)
   n = findall(f, t)
-  ti = t[1:(first(n) - 1)]
-  ts = ntuple(i -> t[n[i]:(n[i + 1] - 1)], length(n) - 1)
-  tf = t[last(n):end]
-  return ti, ts..., tf
+  nsplit = length(n) + 1
+  s = Vector{Any}(undef, nsplit)
+  s[1] = t[1:(first(n) - 1)]
+  for i in 2:(nsplit - 1)
+    s[i] = t[n[i - 1]:(n[i] - 1)]
+  end
+  s[end] = t[last(n):end]
+  return s
 end
+
+## XXX: Very long compile times:
+## https://github.com/JuliaLang/julia/issues/45545
+##
+## julia> using ITensors
+## 
+## julia> @time ITensors.Ops.split(x -> x isa String, ("X", 1))
+##   7.588123 seconds (2.34 M allocations: 100.919 MiB, 1.71% gc time, 100.00% compilation time)
+## ((), ("X", 1))
+## 
+## julia> @time ITensors.Ops.split(x -> x isa String, ("X", 1))
+##   0.042590 seconds (88.59 k allocations: 4.823 MiB, 19.13% gc time, 99.84% compilation time)
+## ((), ("X", 1))
+##
+## function split(f, t::Tuple)
+##   n = findall(f, t)
+##   ti = t[1:(first(n) - 1)]
+##   ts = ntuple(i -> t[n[i]:(n[i + 1] - 1)], length(n) - 1)
+##   tf = t[last(n):end]
+##   return ti, ts..., tf
+## end
 
 struct Op
   which_op
   sites::Tuple
   params::NamedTuple
-  function Op(which_op, sites::Tuple; kwargs...)
-    return new(which_op, sites, NamedTuple(kwargs))
+  function Op(which_op, site...; kwargs...)
+    return new(which_op, site, NamedTuple(kwargs))
   end
-end
-Op(which_op, sites...; kwargs...) = Op(which_op, sites; kwargs...)
-
-function Op(t::Tuple)
-  which_op = first(t)
-  sites_params = Base.tail(t)
-  if last(sites_params) isa NamedTuple
-    sites = Base.front(sites_params)
-    params = last(sites_params)
-  else
-    sites = sites_params
-    params = (;)
-  end
-  return Op(which_op, sites; params...)
 end
 
 which_op(o::Op) = o.which_op
+name(o::Op) = which_op(o)
 sites(o::Op) = o.sites
+site(o::Op) = only(sites(o))
 params(o::Op) = o.params
+
+function (o1::Op == o2::Op)
+  return o1.which_op == o2.which_op && o1.sites == o2.sites && o1.params == o2.params
+end
+
+function hash(o::Op, h::UInt)
+  return hash(which_op(o), hash(sites(o), hash(params(o), hash(:Op, h))))
+end
+
+# Version of `isless` defined for matrices
+_isless(a, b) = isless(a, b)
+_isless(a::AbstractMatrix, b::AbstractMatrix) = isless(hash(a), hash(b))
+_isless(a::AbstractString, b::AbstractMatrix) = true
+_isless(a::AbstractMatrix, b::AbstractString) = !_isless(b, a)
+
+function isless(o1::Op, o2::Op)
+  if sites(o1) ≠ sites(o2)
+    return sites(o1) < sites(o2)
+  end
+  if which_op(o1) ≠ which_op(o2)
+    return _isless(which_op(o1), which_op(o2))
+  end
+  return params(o1) < params(o2)
+end
+
+function isless(o1::Prod{Op}, o2::Prod{Op})
+  if length(o1) ≠ length(o2)
+    return length(o1) < length(o2)
+  end
+  for n in 1:length(o1)
+    if o1[n] ≠ o2[n]
+      return (o1[n] < o2[n])
+    end
+  end
+  return false
+end
+
+function isless(o1::Scaled{C1,Prod{Op}}, o2::Scaled{C2,Prod{Op}}) where {C1,C2}
+  if argument(o1) == argument(o2)
+    if coefficient(o1) ≈ coefficient(o2)
+      return false
+    else
+      c1 = coefficient(o1)
+      c2 = coefficient(o2)
+      #"lexicographic" ordering on  complex numbers
+      return real(c1) < real(c2) || (real(c1) ≈ real(c2) && imag(c1) < imag(c2))
+    end
+  end
+  return argument(o1) < argument(o2)
+end
+
+## function Op(t::Tuple)
+##   which_op = first(t)
+##   site_params = Base.tail(t)
+##   if last(site_params) isa NamedTuple
+##     site = Base.front(site_params)
+##     params = last(site_params)
+##   else
+##     site = site_params
+##     params = (;)
+##   end
+##   return Op(which_op, site; params...)
+## end
+
+## function Op(t::Tuple{WhichOp,NamedTuple,Vararg}) where {WhichOp}
+##   params = t[2]
+##   which_op = t[1]
+##   sites = t[3:end]
+##   return Op(which_op, sites...; params...)
+## end
 
 function sites(a::Union{Sum,Prod})
   s = []
@@ -74,9 +157,15 @@ params(a::Scaled{C,Op}) where {C} = params(argument(a))
 # Op algebra
 #
 
+function convert(::Type{Scaled{C1,Prod{Op}}}, o::Scaled{C2,Prod{Op}}) where {C1,C2}
+  c = convert(C1, coefficient(o))
+  return c * argument(o)
+end
+
 const OpSum{C} = Sum{Scaled{C,Prod{Op}}}
 
-OpSum() = OpSum{Float64}()
+# This helps with in-place operations
+OpSum() = OpSum{ComplexF64}()
 
 (o1::Op + o2::Op) = Applied(sum, ([o1, o2],))
 (o1::Op * o2::Op) = Applied(prod, ([o1, o2],))
@@ -98,6 +187,9 @@ OpSum() = OpSum{Float64}()
 # Op("X", 1) + 1.3 * Op("X", 2)
 (o1::Op + co2::Scaled{C}) where {C} = one(C) * o1 + co2
 
+(o1::Op * o2::Sum) = Applied(sum, (map(a -> o1 * a, only(o2.args)),))
+(o1::Sum * o2::Op) = Applied(sum, (map(a -> a * o2, only(o1.args)),))
+
 # 1.3 * Op("X", 1) + Op("X", 2) * Op("X", 3)
 # 1.3 * Op("X", 1) * Op("X", 2) + Op("X", 3) * Op("X", 4)
 (co1::Scaled{C} + o2::Prod{Op}) where {C} = co1 + one(C) * o2
@@ -116,7 +208,16 @@ adjoint(o::LazyApply.Adjoint{Op}) = only(o.args)
 # Tuple interface
 #
 
-const OpSumLike{C} = Union{Sum{Op},Sum{Scaled{C,Op}},Sum{Prod{Op}},Sum{Scaled{C,Prod{Op}}}}
+const OpSumLike{C} = Union{
+  Sum{Op},
+  Sum{Scaled{C,Op}},
+  Sum{Prod{Op}},
+  Sum{Scaled{C,Prod{Op}}},
+  Prod{Op},
+  Scaled{C,Prod{Op}},
+}
+
+const WhichOp = Union{AbstractString,AbstractMatrix{<:Number}}
 
 # Make a `Scaled{C,Prod{Op}}` from a `Tuple` input,
 # for example:
@@ -128,35 +229,53 @@ function op_term(a::Tuple{Number,Vararg})
   return c * op_term(Base.tail(a))
 end
 
+function op_site(which_op, params::NamedTuple, sites...)
+  return Op(which_op, sites...; params...)
+end
+
+function op_site(which_op, sites_params...)
+  if last(sites_params) isa NamedTuple
+    sites = Base.front(sites_params)
+    params = last(sites_params)
+    return Op(which_op, sites...; params...)
+  end
+  return Op(which_op, sites_params...)
+end
+
 function op_term(a::Tuple{Vararg})
-  a_split = split(x -> x isa AbstractString, a)
+  a_split = split(x -> x isa WhichOp, a)
   @assert isempty(first(a_split))
-  a_split = Base.tail(a_split)
-  o = Op(first(a_split))
-  for aₙ in Base.tail(a_split)
-    o *= Op(aₙ)
+  popfirst!(a_split)
+  o = op_site(first(a_split)...)
+  popfirst!(a_split)
+  for aₙ in a_split
+    o *= op_site(aₙ...)
   end
   return o
 end
 
-function (a1::OpSumLike + a2::Tuple)
-  return a1 + op_term(a2)
+function (o1::OpSumLike + o2::Tuple)
+  return o1 + op_term(o2)
 end
 
-function (a1::OpSumLike - a2::Tuple{Number,Vararg})
-  return a1 + (-first(a2), Base.tail(a2)...)
+function (o1::Tuple + o2::OpSumLike)
+  return op_term(o1) + o2
 end
 
-function (a1::OpSumLike - a2::Tuple{AbstractString,Vararg})
-  return a1 + (-true, a2...)
+function (o1::OpSumLike - o2::Tuple)
+  return o1 - op_term(o2)
 end
 
-function (a1::Prod{Op} * a2::Tuple)
-  return a1 * op_term(a2)
+function (o1::Tuple - o2::OpSumLike)
+  return op_term(o1) - o2
 end
 
-function (a1::Scaled{C,Prod{Op}} * a2::Tuple) where {C}
-  return a1 * op_term(a2)
+function (o1::OpSumLike * o2::Tuple)
+  return o1 * op_term(o2)
+end
+
+function (o1::Tuple * o2::OpSumLike)
+  return op_term(o1) * o2
 end
 
 function show(io::IO, ::MIME"text/plain", o::Op)
@@ -170,15 +289,22 @@ end
 show(io::IO, o::Op) = show(io, MIME("text/plain"), o)
 
 function show(io::IO, ::MIME"text/plain", o::Prod{Op})
-  for oₙ in o
-    print(io, oₙ, " ")
+  for n in 1:length(o)
+    print(io, o[n])
+    if n < length(o)
+      print(io, " ")
+    end
   end
   return nothing
 end
 show(io::IO, o::Prod{Op}) where {C} = show(io, MIME("text/plain"), o)
 
-function show(io::IO, m::MIME"text/plain", o::Scaled{C,Prod{Op}}) where {C}
-  print(io, coefficient(o))
+function show(io::IO, m::MIME"text/plain", o::Scaled{C,O}) where {C,O<:Union{Op,Prod{Op}}}
+  c = coefficient(o)
+  if isreal(c)
+    c = real(c)
+  end
+  print(io, c)
   print(io, " ")
   show(io, m, argument(o))
   return nothing