+++ src/mps/abstractmps.jl
@@ -1084,6 +1084,9 @@ function _log_or_not_dot(
   end
 
   if loginner
+    if !isreal(O[]) || real(O[]) < 0
+      log_inner_tot += log(complex(O[]))
+    end
     return log_inner_tot
   end
 
@@ -1213,7 +1216,7 @@ function lognorm(M::AbstractMPS)
       "log(norm²) is $lognorm2_M, which is not real up to a relative tolerance of $rtol"
     )
   end
-  return 0.5 * lognorm2_M
+  return 0.5 * real(lognorm2_M)
 end
 
 function isapprox(
@@ -1378,9 +1381,9 @@ println()
       inner(ψ₃, ψ₁) + 2 * inner(ψ₃, ψ₂) + inner(ψ₃, ψ₃)
 ```
 """
-function +(ψ⃗::MPST...; cutoff=1e-15, kwargs...) where {MPST<:AbstractMPS}
-  # TODO: Check that the inputs have the same site indices
-
+function +(
+  ::Algorithm"densitymatrix", ψ⃗::MPST...; cutoff=1e-15, kwargs...
+) where {MPST<:AbstractMPS}
   if !all(ψ -> hassameinds(siteinds, first(ψ⃗), ψ), ψ⃗)
     error(
       "In `+(::MPS/MPO...)`, the input `MPS` or `MPO` do not have the same site indices. For example, the site indices of the first site are $(siteinds.(ψ⃗, 1))",
@@ -1446,6 +1449,46 @@ function +(ψ⃗::MPST...; cutoff=1e-15, kwargs...) where {MPST<:AbstractMPS}
   set_ortho_lims!(ψ, 1:1)
 
   return convert(MPST, ψ)
+end
+
+function +(::Algorithm"directsum", ψ⃗::MPST...) where {MPST<:AbstractMPS}
+  n = length(first(ψ⃗))
+  @assert all(ψᵢ -> length(first(ψ⃗)) == length(ψᵢ), ψ⃗)
+
+  # Output tensor
+  ϕ = MPS(n)
+
+  # Direct sum first tensor
+  j = 1
+  l⃗j = map(ψᵢ -> linkind(ψᵢ, j), ψ⃗)
+  ϕj, (lj,) = directsum(
+    (ψ⃗[i][j] => (l⃗j[i],) for i in 1:length(ψ⃗))...; tags=[tags(first(l⃗j))]
+  )
+  ljm_prev = lj
+  ϕ[j] = ϕj
+  for j in 2:(n - 1)
+    l⃗jm = map(ψᵢ -> linkind(ψᵢ, j - 1), ψ⃗)
+    l⃗j = map(ψᵢ -> linkind(ψᵢ, j), ψ⃗)
+    ϕj, (ljm, lj) = directsum(
+      (ψ⃗[i][j] => (l⃗jm[i], l⃗j[i]) for i in 1:length(ψ⃗))...;
+      tags=[tags(first(l⃗jm)), tags(first(l⃗j))],
+    )
+    ϕj = replaceind(ϕj, ljm => dag(ljm_prev))
+    ljm_prev = lj
+    ϕ[j] = ϕj
+  end
+  j = n
+  l⃗jm = map(ψᵢ -> linkind(ψᵢ, j - 1), ψ⃗)
+  ϕj, (ljm,) = directsum(
+    (ψ⃗[i][j] => (l⃗jm[i],) for i in 1:length(ψ⃗))...; tags=[tags(first(l⃗jm))]
+  )
+  ϕj = replaceind(ϕj, ljm => dag(ljm_prev))
+  ϕ[j] = ϕj
+  return ϕ
+end
+
+function +(ψ⃗::AbstractMPS...; alg=Algorithm"densitymatrix"(), kwargs...)
+  return +(Algorithm(alg), ψ⃗...; kwargs...)
 end
 
 +(ψ::AbstractMPS) = ψ
@@ -2106,7 +2149,7 @@ expτH = ops(os, s)
 ```
 """
 function product(
-  As::Vector{<:ITensor},
+  As::Vector{ITensor},
   ψ::AbstractMPS;
   move_sites_back_between_gates::Bool=true,
   move_sites_back::Bool=true,
@@ -2123,6 +2166,27 @@ function product(
     Aψ = movesites(Aψ, ns .=> ñs; kwargs...)
   end
   return Aψ
+end
+
+# Apply in the reverse order for proper order of operations
+# For example:
+#
+# s = siteinds("Qubit", 1)
+# ψ = randomMPS(s)
+#
+# # U = Z₁X₁
+# U = Prod{Op}()
+# U = ("X", 1) * U
+# U = ("Z", 1) * U
+#
+# # U|ψ⟩ = Z₁X₁|ψ⟩
+# apply(U, 
+function product(o::Prod{ITensor}, ψ::AbstractMPS; kwargs...)
+  return product(reverse(terms(o)), ψ; kwargs...)
+end
+
+function (o::Prod{ITensor})(ψ::AbstractMPS; kwargs...)
+  return apply(o, ψ; kwargs...)
 end
 
 #