+++ test/mps.jl
@@ -197,6 +197,17 @@ include("util.jl")
     @test norm(phic[4]) ≈ 1.0
   end
 
+  @testset "randomMPS with chi>1" for linkdims in [1, 4]
+    phi = randomMPS(Float32, sites; linkdims)
+    @test LinearAlgebra.promote_leaf_eltypes(phi) === Float32
+    @test all(x -> eltype(x) === Float32, phi)
+    @test maxlinkdim(phi) == linkdims
+    phic = randomMPS(ComplexF32, sites; linkdims)
+    @test LinearAlgebra.promote_leaf_eltypes(phic) === ComplexF32
+    @test maxlinkdim(phic) == linkdims
+    @test all(x -> eltype(x) === ComplexF32, phic)
+  end
+
   @testset "inner different MPS" begin
     phi = randomMPS(sites)
     psi = randomMPS(sites)
@@ -209,6 +220,19 @@ include("util.jl")
     badsites = [Index(2) for n in 1:(length(psi) + 1)]
     badpsi = randomMPS(badsites)
     @test_throws DimensionMismatch inner(phi, badpsi)
+  end
+
+  @testset "loginner" begin
+    n = 4
+    c = 2
+
+    s = siteinds("S=1/2", n)
+    ψ = c .* randomMPS(s; linkdims=4)
+    @test exp(loginner(ψ, ψ)) ≈ c^(2n)
+    @test exp(loginner(ψ, -ψ)) ≈ -c^(2n)
+
+    α = randn(ComplexF64)
+    @test exp(loginner(ψ, α * ψ)) ≈ α * c^(2n)
   end
 
   @testset "broadcasting" begin
@@ -444,6 +468,23 @@ include("util.jl")
     K12 = Ks[1] + Ks[2]
     K123 = K12 + Ks[3]
     @test inner(sum(Ks), K123) ≈ inner(K123, K123)
+
+    χ1 = 2
+    χ2 = 3
+    ψ1 = randomMPS(sites; linkdims=χ1)
+    ψ2 = 0.0 * randomMPS(sites; linkdims=χ2)
+
+    ϕ1 = +(ψ1, ψ2; alg="densitymatrix", cutoff=nothing)
+    for j in 2:7
+      @test linkdim(ϕ1, j) == χ1 + χ2
+    end
+    @test inner(ϕ1, ψ1) + inner(ϕ1, ψ2) ≈ inner(ϕ1, ϕ1)
+
+    ϕ2 = +(ψ1, ψ2; alg="directsum")
+    for j in 1:8
+      @test linkdim(ϕ2, j) == χ1 + χ2
+    end
+    @test inner(ϕ2, ψ1) + inner(ϕ2, ψ2) ≈ inner(ϕ2, ϕ2)
   end
 
   @testset "+ MPS with coefficients" begin