+++ ITensorGaussianMPS/NEWS.md
@@ -6,6 +6,16 @@ Note that as of Julia v1.5, in order to see deprecation warnings you will need t
 
 After we release v1 of the package, we will start following [semantic versioning](https://semver.org).
 
+ITensors v0.0.4 Release Notes
+=============================
+
+Bugs:
+
+Enhancements:
+
+- Update for new OpSum representation and interface (#920)
+- Add check for proper fermionic operators when making hopping Hamiltonian (#920)
+
 ITensors v0.0.3 Release Notes
 =============================